pyproject.toml
[project]
name = "api-final-yatube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==3.2.16",
    "django-filter==22.1",
    "djangorestframework==3.12.4",
    "djangorestframework-simplejwt==4.7.2",
    "djoser==2.1.0",
    "pillow==9.3.0",
    "pyjwt==2.1.0",
    "pytest==6.2.4",
    "pytest-django==4.4.0",
    "pytest-pythonpath==0.7.3",
    "requests==2.26.0",
    "setuptools>=80.9.0",
]

[tool.ruff]
line-length = 78
